crashinfo types
